Some snp.index updates and new Logos (#101)
* Some Updates to snp.index for Cable Vodafone

* New Logo mdr (SD & HD)

All .svg are saved with Inkscape like in the wiki

* New Logo Disney Channel (SD & HD)

All .svg are saved with Inkscape like in the wiki

* New Logos Sonnenklar TV (SD & HD)

I replace the default Logos also. the first reason is they wasn't optimized. the second ...so all files are based on the same basic structure.

All .svg are saved with Inkscape like in the wiki

* Create nicetv.white.png